[device-minion-runtime] [SW-7031]: Updating head, tail and length values to be 64-bit in circular buffer. Made sure that the source and destination addresses for head and tail pointers are 64-bit aligned. Added fallback mechanism for tail mismatch.

+++ device-minion-runtime/src/shared/include/etsoc_memory.h
@@ -32,67 +32,67 @@ enum ETSOC_MEM_TYPES
 };
 
 /*********************/
-/*! \fn void ETSOC_Memory_Read_Uncacheable(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Read_Uncacheable(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Reads data from ETSOC uncacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be read.
     \returns None.
 */
-void ETSOC_Memory_Read_Uncacheable(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Read_Uncacheable(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Write_Uncacheable(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Write_Uncacheable(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Writes data to ETSOC uncacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be written.
     \returns None.
 */
-void ETSOC_Memory_Write_Uncacheable(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Write_Uncacheable(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Read_Local_Atomic(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Read_Local_Atomic(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Reads data from ETSOC L2 cacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be read.
     \returns None.
 */
-void ETSOC_Memory_Read_Local_Atomic(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Read_Local_Atomic(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Write_Local_Atomic(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Write_Local_Atomic(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Writes data to ETSOC L2 cacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be written.
     \returns None.
 */
-void ETSOC_Memory_Write_Local_Atomic(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Write_Local_Atomic(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Read_Global_Atomic(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Read_Global_Atomic(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Reads data from ETSOC L3 cacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be read.
     \returns None.
 */
-void ETSOC_Memory_Read_Global_Atomic(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Read_Global_Atomic(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Write_Global_Atomic(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Write_Global_Atomic(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Writes data to ETSOC L3 cacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be written.
     \returns None.
 */
-void ETSOC_Memory_Write_Global_Atomic(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Write_Global_Atomic(const void *src_ptr, void *dest_ptr, uint64_t length);
 
-/*! \fn void ETSOC_Memory_Read_Write_Cacheable(void *src_ptr, void *dest_ptr, uint32_t length)
+/*! \fn void ETSOC_Memory_Read_Write_Cacheable(void *src_ptr, void *dest_ptr, uint64_t length)
     \brief Reads/Writes data from/to ETSOC cacheable memory.
     \param [in] src_ptr: Pointer to source data buffer.
     \param [in] dest_ptr: Pointer to destination data buffer.
     \param [in] length: Total length (in bytes) of the data that needs to be read/written.
     \returns None.
 */
-void ETSOC_Memory_Read_Write_Cacheable(const void *src_ptr, void *dest_ptr, uint32_t length);
+void ETSOC_Memory_Read_Write_Cacheable(const void *src_ptr, void *dest_ptr, uint64_t length);
 
 #endif /* ETSOC_MEMORY_DEFS_H_ */
